qcacmn: Add macros to fetch the phymode from channel
Add channel macros to query if an input channel is 11ac/11ax/11be.

CRs-Fixed: 3202132

+++ umac/global_umac_dispatcher/lmac_if/inc/wlan_lmac_if_def.h
@@ -2144,7 +2144,7 @@ struct wlan_lmac_if_dfs_rx_ops {
 	QDF_STATUS (*dfs_set_postnol_freq)(struct wlan_objmgr_pdev *pdev,
 					   qdf_freq_t postnol_freq);
 	QDF_STATUS (*dfs_set_postnol_mode)(struct wlan_objmgr_pdev *pdev,
-					   uint8_t postnol_mode);
+					   uint16_t postnol_mode);
 	QDF_STATUS (*dfs_set_postnol_cfreq2)(struct wlan_objmgr_pdev *pdev,
 					     qdf_freq_t postnol_cfreq2);
 	QDF_STATUS (*dfs_get_postnol_freq)(struct wlan_objmgr_pdev *pdev,
